Refactor: move Navbar and Footer to App.jsx for proper Separation of Concerns
